% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count_Unique_Childrens_IDs.R
\name{CountuniqueChildren}
\alias{CountuniqueChildren}
\title{Count Unique Children's IDs in a Dataframe}
\usage{
CountuniqueChildren(dat_and_col)
}
\arguments{
\item{dat_and_col}{Name of the Dataframe and column of the IDs, in the format dat$ID}
}
\value{
The number of unique children's IDs, ending in "-01".
}
\description{
Count Unique Children's IDs in a Dataframe
}
\examples{
my_data<-data.frame(aliquot=c("GH56543", "KJ98757", "BW98564"),
ID=c("MAD-5-16-78-00",  "MAD-5-17-325-00", "MAD-5-18-199-01"))
CountuniqueChildren(my_data$ID)
}
